requirements.txt
unsloth
torch
transformers>=4.46.0
datasets>=2.21.0
peft
bitsandbytes
accelerate
pandas
scikit-learn
pyyaml
matplotlib
seaborn
tqdm
sentencepiece
protobuf
